backend/requirements.txt
asn1crypto==0.24.0
astroid==2.1.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.5
dj-database-url==0.5.0
Django==2.1.5
django-cors-headers==2.4.0
djangorestframework==3.9.1
gunicorn==19.9.0
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2==2.7.7
pycparser==2.19
PyJWT==1.7.1
pylint==2.2.2
pytz==2018.9
requests==2.21.0
six==1.12.0
spotipy==2.4.4
urllib3==1.24.1
wrapt==1.11.1
